Support for multiple Hellas special prefixes and fix of uppercase prefixes
Generalised functionality that allowed for a special prefix for underdot to allow
for a few different special prefixes which are to be treated the same way. Now the only one added
is macron, apart from underdot.

Also, uppercase prefixes for diacritics in hellas are different from lowercase, this was corrected




git-svn-id: svn+ssh://svn.code.sf.net/p/epidoc/code/trunk@1979 85be4e97-ed4a-0410-b6a7-8d1dfdb05c0b